+++ packages/base/src/utils/warnOnce.ts
@@ -1,0 +1,5 @@
+import once from './private/once';
+
+export default function warnOnce(message: string): (...args: any[]) => void {
+  return once((...args: any[]) => console.warn(`botframework-webchat: ${message}`, ...args));
+}