+++ packages/bundle/src/boot/exports/index.ts
@@ -14,7 +14,23 @@ export {
 // #endregion
 
 // #region Overrides for backward compatibility
-// Web devs should move to named exports instead.
-export * as Components from '../actual/component/full';
-export * as hooks from '../actual/hook/full';
+import { deprecateObject } from '@msinternal/botframework-webchat-base/utils';
+
+import * as hooks from '../actual/hook/full';
+
+const deprecatedHooks = deprecateObject(
+  hooks,
+  "`import { hooks } from 'botframework-webchat'` has been deprecated, use `import { %s } from 'botframework-webchat/hook'` instead."
+) as typeof hooks; // Need explicit type casting.
+
+export { deprecatedHooks as hooks };
+
+import * as Components from '../actual/component/full';
+
+const deprecatedComponents = deprecateObject(
+  Components,
+  "`import { Components } from 'botframework-webchat'` has been deprecated, use `import { %s } from 'botframework-webchat/component'` instead."
+) as typeof Components; // Need explicit type casting.
+
+export { deprecatedComponents as Components };
 // #endregion